bst/SplayTree: correct rotation algorithm
The rotation algorithm can be confusing as we make use of an explicit stack
instead of recursion.

During insertion, we start from the root and push each node we traversed onto
a stack, and the direction of the pointer we followed is pushed onto another
stack. This direction is that of the child of the parent node.

When we finally hit a null node, the direction of this node is already on the
stack. As such, it is a mistake to push the direction onto the stack again.
This is corrected in this patch.

In addition, we confused what each element in the direction stack represents.
During the rotation up to root, we actually make use of the direction from:

grandparent       -> parent
great grandparent -> parent

instead of:

parent      -> current node
grandparent -> parent

which is also corrected here.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>