parser: suppress printing of result for invalid computation
When there is a division by 0 somewhere inside the binary tree, we propagate
the error upwards in bst_postorder_compute so that the computation stops.
This results in printing of the following message:

Error in computation

However, we go on to print the result of the invalid computation as 0 when
nothing should be printed at all.

Suppress the printing of the result for an invalid computation since it is
undefined.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>