Allow user to indicate destination file
We made use of the "compressed" and "decompressed" files for files which we
compress to and files which we decompress to respectively. Now that our
argument parser is up, allow the user to indicate a destination file he wants
the output to go to.

If the user does not supply the output file names, we will use the default
output filenames.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>